Fix event spying example in README
Passing the event spy to `expect()` works when verifying a general
behavior or result, such as with `toHaveBeenTriggered()`. However,it
seems like the event name should be used instead when verifying behavior
against an element, such as with `toHaveBeenTriggeredOn()`.

For reference:
https://github.com/velesin/jasmine-jquery#event-spies

This also corrects a typo with a missing comma in the example.